requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
Pillow>=11.1.0
aiohttp>=3.9.3
asyncio>=3.4.3
beautifulsoup4==4.12.2 